modified:   client/src/component/AppRoutes.jsx     adding route /login to LoginForm 	modified:   client/src/component/NavBar/NavBar.jsx     adding login link to navbar 	new file:   client/src/component/userForm/LoginForm.jsx     rename login form 	deleted:    client/src/component/userForm/LoginModal.jsx     delete login modal
