+++ client/src/component/AppRoutes.jsx
@@ -4,6 +4,7 @@ import ShopDetails from "./Shops/ShopDetails";
 import ShopsList from "./Shops/ShopsList";
 import NewProductForm from "./products/NewProductForm";
 import ProductDetails from "./products/ProductDetails";
+import LoginForm from "./userForm/LoginForm";
 
 function AppRoutes() {
   return (
@@ -13,6 +14,7 @@ function AppRoutes() {
       <Route path="shops/new" element={<NewShopForm />} />
       <Route path="shops/:id/products/new" element={<NewProductForm />} />
       <Route path="shops/:id/products/:id" element={<ProductDetails />} />
+      <Route path="/login" element={<LoginForm />} />
     </Routes>
   );
 }
